Complete connection setup portion of `TransportHandler`
